Changed various unnecessary BigInts to ints. Finished parseObject().

+++ src/org/distroverse/dvtp/DvtpExternalizable.java
@@ -15,5 +15,5 @@ import java.io.Externalizable;
  */
 public interface DvtpExternalizable extends Externalizable
    {
-   public abstract BigInt getClassNumber();
+   public abstract int getClassNumber();
    }
